requirements.txt
geobr
streamlit
geopandas
pandas
folium
plotly.express
numpy
streamlit-folium
seaborn
streamlit-aggrid
wordcloud